Настроил focus на TextField с помощью inputRef={inputRef}
